Fix: quote flymake--mode-line-format symbol boundp check
It seems like the symbol flymake--mode-line-format is no longer found
in the flymake package's code, so this was raising a void symbol
error. Quoting it should keep it backwards compatible for whatever
behavior this was using previously, and prevent the error from making
it so the flymake-mode-line-format variable was not being processed.